+++ Assets/CommonCore/Core/Config/ConfigState.cs
@@ -38,8 +38,11 @@ namespace CommonCore.Config
                 if(didMigrate)
                 {
                     Debug.Log("[Config] Config file was migrated successfully");
-                    Directory.CreateDirectory(System.IO.Path.Combine(CoreParams.PersistentDataPath, "migrationbackups"));
-                    File.Copy(Path, System.IO.Path.Combine(CoreParams.PersistentDataPath, "migrationbackups" , $"config.migrated.{DateTime.Now.ToString("yyyy-MM-dd_HHmmss")}.json"), true);
+                    if(CoreParams.UseMigrationBackups || CoreParams.IsDebug)
+                    {
+                        Directory.CreateDirectory(System.IO.Path.Combine(CoreParams.PersistentDataPath, "migrationbackups"));
+                        File.Copy(Path, System.IO.Path.Combine(CoreParams.PersistentDataPath, "migrationbackups", $"config.migrated.{DateTime.Now.ToString("yyyy-MM-dd_HHmmss")}.json"), true);
+                    }                    
                     CoreUtils.WriteExternalJson(Path, rawConfig);
                 }
                 Instance = CoreUtils.InterpretJson<ConfigState>(newRawConfig ?? rawConfig);